pyproject.toml
[project]
name = "quotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-basicauth>=0.2.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "ruff>=0.12.2",
]

[tool.ruff]
fix = true
line-length = 120

[tool.ruff.lint]
select = ["I"]

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"
line-ending = "lf"
